Limit DisasmIter to go1.7 and onward
